refactor: address daily quality findings for 2026-08-11
Split Token Usage overview tab and Code Agent Behaviour settings into
dedicated modules; centralize pending Linear link storage and Task Linear
page-size constants; archive quality score 75 report.
