+++ apps/web/src/features/task/lib/task-linear-panel-model.ts
@@ -1,0 +1,3 @@
+/** Pure helpers and constants for the Task Linear list panel. */
+
+export const TASK_LINEAR_PAGE_SIZE = 20;
